+++ src/components/AddRestaurant.ts
@@ -1,7 +1,5 @@
-import Validator from "../tools/Validator";
-import IRestaurant from "../type/IRestaurant";
-import BottomSheet from "./BottomSheet";
-import RestaurantList from "./RestaurantList";
+import { closeBottomSheet } from "../domain/bottomSheet";
+import { createNewRestaurant, tryAddNewRestaurant } from "../domain/restaurant";
 
 class AddRestaurant extends HTMLElement {
   constructor() {
@@ -66,17 +64,10 @@ class AddRestaurant extends HTMLElement {
       `;
   }
 
-  closeBottomSheet() {
-    const bottomSheet = document.getElementById("bottomSheet");
-    if (bottomSheet instanceof BottomSheet) {
-      bottomSheet.close();
-    }
-  }
-
   onClickCancelButton() {
     const cancelButton = document.getElementById("cancelButton");
     cancelButton?.addEventListener("click", () => {
-      this.closeBottomSheet();
+      closeBottomSheet();
     });
   }
 
@@ -84,36 +75,8 @@ class AddRestaurant extends HTMLElement {
     const restaurantForm = document.getElementById("restaurantForm");
     restaurantForm?.addEventListener("submit", (event) => {
       event.preventDefault();
-      try {
-        const newRestaurant = this.createNewRestaurant(event);
-        Validator.checkRestaurant(newRestaurant);
-        this.addNewRestaurant(newRestaurant);
-        this.closeBottomSheet();
-      } catch (error) {
-        if (error instanceof Error) {
-          alert(error.message);
-        }
-      }
+      tryAddNewRestaurant(createNewRestaurant(event));
     });
-  }
-
-  addNewRestaurant(newRestaurant: IRestaurant) {
-    const restaurantList = document.getElementById("restaurantList");
-    if (restaurantList instanceof RestaurantList) {
-      restaurantList.addRestaurant(newRestaurant);
-    }
-  }
-
-  createNewRestaurant(event: SubmitEvent) {
-    const formData = new FormData(event.target as HTMLFormElement);
-    const newRestaurant: IRestaurant = {
-      category: formData.get("category") as string,
-      name: formData.get("name") as string,
-      distance: Number(formData.get("distance")),
-      description: formData.get("description") as string,
-      link: formData.get("link") as string,
-    };
-    return newRestaurant;
   }
 }
